requirements.txt
flask
flask-cors
openai-whisper
deep-translator
langdetect
torch
torchaudio
ffmpeg-python

